This Won't Work
Current issue is that blender can't access context on addon startup because it'll do fuck all since it really has to be run when that portion of script is run -- So I need to find a workaround -- https://stackoverflow.com/questions/47064223/blender-attributeerror-restrictcontext-object-has-no-attribute-scene --  " Restricted Context

To avoid errors on accessing blend-file data while addons register() / unregister(), Blender now restricts access to bpy.context and bpy.data.

This is done because there is no assurance that the data loaded when the addon is registered will be active or even exist when the user accesses operators the addon defines. "